Fix model (polyPoint fields + addedge) + added fix in controller
Nu kan polypoints endelig få om der kan køre en cykel, gå, speedlimit. Det fungerede ikke før men nu gør det.
+ fix for dijkstra

+++ Map of Denmark/app/src/main/java/bfst22/vector/Graph.java
@@ -104,7 +104,7 @@ public class Graph {
 
         public float setWeightDistance (PolyPoint from, PolyPoint to,float speedlimit){
             Distance d = new Distance();
-            return (float)d.haversineFormula(from, to) / speedlimit;
+            return d.haversineFormula(from, to) / speedlimit;
         }
 
         public float setWeight(float weight){
